+++ mail/mail-common/src/mail_scroller.rs
@@ -352,12 +352,21 @@ pub struct ScrollerWorker<T: MailScrollerSource + 'static> {
     ctx: Weak<MailUserContext>,
     source: Arc<RwLock<T>>,
     task: MailPaginatorJoinHandle,
+    execute_on_online: Option<AbortHandle>,
     items: Vec<T::Item>,
     page_size: usize,
     previous_update: Option<ScrollerSource>,
     update: flume::Sender<ScrollerUpdate<T::Item>>,
     ordered_command_recv: flume::Receiver<ScrollerOrderedCommand>,
     ordered_command_send: flume::Sender<ScrollerOrderedCommand>,
+}
+
+impl<T: MailScrollerSource + 'static> Drop for ScrollerWorker<T> {
+    fn drop(&mut self) {
+        if let Some(handle) = self.execute_on_online.take() {
+            handle.abort();
+        }
+    }
 }
 
 impl<T: MailScrollerSource + 'static> ScrollerWorker<T> {
@@ -387,6 +396,7 @@ impl<T: MailScrollerSource + 'static> ScrollerWorker<T> {
             source,
             page_size,
             task,
+            execute_on_online: None,
             items: vec![],
             previous_update: None,
             update: update_sender,
@@ -637,6 +647,19 @@ impl<T: MailScrollerSource + 'static> ScrollerWorker<T> {
 
         if items.is_empty() && has_more_in_label {
             if self.task.is_none() {
+                if self.execute_on_online.is_none() {
+                    let ctx_clone = ctx.clone();
+                    let channel = self.ordered_command_send.clone();
+                    let handle = ctx.spawn(async move {
+                        ctx_clone
+                            .network_monitor_service()
+                            .network_status_observer()
+                            .wait_until_online()
+                            .await;
+                        Self::schedule_fetch_more(&channel, call_src).await;
+                    });
+                    self.execute_on_online = Some(handle.abort_handle());
+                }
                 // We will not progress any further without task,
                 // and task will be spawned only when we are online,
                 // lets wait for another call.
@@ -646,11 +669,12 @@ impl<T: MailScrollerSource + 'static> ScrollerWorker<T> {
                 // if we requested it once with no effect.
                 self.previous_update = None;
                 tracing::debug!("No items to return, requesting additional fetch more");
-                self.ordered_command_send
-                    .send_async(ScrollerOrderedCommand::FetchMore(call_src))
-                    .await
-                    .map_err(|e| anyhow!("Failed to schedule fetch more command: {e:?}"))?;
+                Self::schedule_fetch_more(&self.ordered_command_send, call_src).await;
             }
+        }
+
+        if let Some(handle) = self.execute_on_online.take() {
+            handle.abort();
         }
 
         if items.is_empty() {
@@ -659,6 +683,7 @@ impl<T: MailScrollerSource + 'static> ScrollerWorker<T> {
         } else {
             self.previous_update = Some(call_src);
             tracing::debug!("New items fetched: {}", items.len());
+            self.previous_update = Some(call_src);
             self.items.extend(items.clone());
             Ok(ScrollerUpdate::Append {
                 src: call_src,
@@ -788,11 +813,7 @@ impl<T: MailScrollerSource + 'static> ScrollerWorker<T> {
 
         if cant_see_first_page {
             tracing::info!("We do not see the first page, requesting fetch more");
-            let _ = self
-                .ordered_command_send
-                .send_async(ScrollerOrderedCommand::FetchMore(src))
-                .await
-                .inspect_err(|e| tracing::error!("Failed to send append update: {e:?}"));
+            Self::schedule_fetch_more(&self.ordered_command_send, src).await;
         }
 
         Ok(())
@@ -818,7 +839,7 @@ impl<T: MailScrollerSource + 'static> ScrollerWorker<T> {
         self.task = task;
 
         if items.is_empty() && self.task.is_none() {
-            let status = ctx.network_monitor_service().status();
+            let status = ctx.network_monitor_service().combined_status();
             tracing::warn!(
                 "No items and no task to return - status: {status:?}, previous fetch result: {result:?}"
             );
@@ -830,7 +851,7 @@ impl<T: MailScrollerSource + 'static> ScrollerWorker<T> {
 
     async fn wait_for_request(&mut self) -> Result<(), MailContextError> {
         let ctx = self.ctx.upgrade().ok_or(MailContextError::MissingContext)?;
-        let is_online = ctx.network_monitor_service().status().is_online();
+        let is_online = ctx.network_monitor_service().is_os_online();
 
         if self.task.is_some() && is_online {
             Self::await_task(&mut self.task).await
@@ -850,6 +871,16 @@ impl<T: MailScrollerSource + 'static> ScrollerWorker<T> {
         } else {
             Ok(())
         }
+    }
+
+    async fn schedule_fetch_more(
+        channel: &flume::Sender<ScrollerOrderedCommand>,
+        src: ScrollerSource,
+    ) {
+        let _ = channel
+            .send_async(ScrollerOrderedCommand::FetchMore(src))
+            .await
+            .inspect_err(|e| tracing::error!("Failed to schedule fetch more command: {e:?}"));
     }
 }