+++ mail/mail-common/src/models/message.rs
@@ -1261,7 +1261,7 @@ impl Message {
             return Err(AppError::MessageHasNoRemoteId(self.id()).into());
         };
 
-        if with_network_check && ctx.network_monitor_service().check_now().await.is_offline() {
+        if with_network_check && ctx.network_monitor_service().is_os_offline() {
             debug!("No connection, skipping sync");
             return Err(MailContextError::Api(ApiServiceError::NetworkError(
                 "No connection".to_owned(),