+++ mail/mail-uniffi/CHANGELOG.md
@@ -5,7 +5,39 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [mail-uniffi-v0.142.9] - 2025-09-01
+
+### Fixes
+
+- Scroller is allowed to schedule automatic `fetch_more` in certain circumstances
+
+
+## [mail-uniffi-v0.142.8] - 2025-08-29
+
+### Fixes
+
+- Run immediate checks in the background
+
+
+## [mail-uniffi-v0.142.7] - 2025-08-29
+
+### Fixes
+
+- Prevent looping on false client status by checking the combined status
+
+
+## [mail-uniffi-v0.142.6] - 2025-08-29
+
+### Changed
+
+- Expose separate os level network observer
+- Check network status on resume_work/on_enter_foreground
+- Conversation and message body fetch check OS network status
+- Integrate OS level status checks in the scroller
+- Scroller schedules fetch_more automatically when offline to be executed once online
+
+
+## [mail-uniffi-v0.142.5] - 2025-08-29
 
 ### Changed